[ios clean] Disables ToolsMenu items based on ToolsMenuConfig
This CL disables certain items on the ToolsMenu based on the following 
criteria:
-No Tabs Opened. (Will disable Close All Tabs/Incognito Tabs)
-In New Tab Page. (Will disable Find in Page and Request Desktop/Mobile
sites).

In order to do this a new Bitmask is added to the ToolsMenuModel and it
is used by the ToolsMediator to check if the item should be enabled or
not.

Bug: 682880
Change-Id: I2318f7c3d6308317cb9c07e6b5a4692a4fdd9a5e
Reviewed-on: https://chromium-review.googlesource.com/571922
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#488442}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0825547dd2acbe7aea70481c992c4540ad918fbf