Increase timeout in WebStateTest.Snapshot
The current timeout for page load (0.1s) is a bit short and leads to
flakyness on official bots on iPad Air.
I can reproduce the flakyness locally.
Increase the timeout to 1s as the loading is not the tested part.

Bug: None
Change-Id: Ie944a99590f8083e2c22d43769dcdf7d9a0dcab1
Reviewed-on: https://chromium-review.googlesource.com/579369
Reviewed-by: Ed Chin <edchin@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#488649}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d566194dd15fd26180f7bb94fe060d978535953b