Create MetricsRecorder protocol.
Create a MetricsRecorder protocol that allows implementors to record
a metric for a given NSInvocation. Objects implementing this protocol
will be able to register with the dispatcher to be called when a
relevant selector is invoked on the dispatcher.

Bug: none
Change-Id: Icb0074f6d3857ea9b7c0043c052e5e35dc6598ec
Reviewed-on: https://chromium-review.googlesource.com/578305
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Gregory Chatzinoff <gchatz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#488710}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 70bac5d90548b60766f465c6dfed206b32a38ebd