Revert of "[Sync] Maintain a global_id mapping to update UserEvents that references navigations."
Reverts patch set 6 of https://codereview.chromium.org/2958303002

Reason for revert:
Causes crashes on iOS. See https://crbug.com/747658.

TBR=skym@chromium.org,pnoland@chromium.org,holte@chromium.org

Original issue's description:

UserEvents are created and reference sessions data by global_id. It is
important that they travel over the wire in this format, as this is how
the sessions data is keyed into Footprints, and the sync server cannot
feasibly perform a persistent read before writing new UserEvents.
Because global_id is forced to be part of the wire interface and
specifics, it is also conveniently part of the interface to the
UserEventService.

However, while this global_id is seemingly sacred to sync, it comes
from simply the timestmap of the navigation entry. And if a page is
reloaded, this value can change. A problem arises when a UserEvent
references a navigation by a global_id that was updated and the old
value is never going to reach the sync server.

To fix we maintain two sets of mappings, one that allows us to map old
global_ids to new global_id, and one that allows us to efficiently look
up uncommitted UserEvents by global_id, so that if we get either a new
global_id or a new UserEvent we can quickly fix the global_id.

BUG=747621

Review-Url: https://codereview.chromium.org/2958303002
Cr-Original-Commit-Position: refs/heads/master@{#488944}
Committed: https://chromium.googlesource.com/chromium/src/+/a58a1f014663671847efd294313b99fe71e1a6bf
Change-Id: I109949b56fa028574b53444285b6f83305367560
Reviewed-on: https://chromium-review.googlesource.com/581689
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Original-Commit-Position: refs/heads/master@{#488826}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3105b147ea6a4178791a717a375c4a4ad57e489c