Include additional supported referrer meta tag values.
Follow up CL for https://chromium-review.googlesource.com/c/580630/

When setting the referrer policy to an unsupported value using a meta
tag, the console for WKWebView prints out a list of accepted values,
which was used to create the initial version of this CL (linked above).
However, there are additional values (i.e. 'default') that are supported
yet not listed in the console message.  This CL adds this value and
fixes a typo in the previous list so that the context menu JS referrer
policy matches WKWebView's implementation.

Bug: 733202
Change-Id: Ied5622d53c4110f4dad175fdb3deb2eabed507e6
Reviewed-on: https://chromium-review.googlesource.com/583717
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#489135}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 89dffe9216c94e93de5741a520c67ccce7aa5675