[iOS] Add support for per-WebState Mojo interfaces
//content-based platforms have the concept of per-frame interfaces
(i.e., interfaces that are scoped to the connection between a render
frame and its corresponding host). These interfaces replace the per-frame
IPC between the renderer and browser that was used to implement a given
feature (e.g., translate). Some of these interfaces will in the
long term be shared with iOS as the process of servicification unfolds,
e.g. for features like translate and autofill. In this long-term picture,
the browser-side implementation of the Mojo interface will be shared
between iOS and other platforms, with a platform-specific client
invoking this interface via Mojo (from the renderer process on !iOS,
and likely originating from Javascript on iOS).

To facilitate this long-term picture, this CL adds corresponding
support for per-WebState interfaces on iOS. Note that while the
infrastructure is added at the RenderFrame(Host) granularity in
//content, on iOS all communication is implicitly occurring within the
context of the main frame, and there is thus no equivalent concept to
RenderFrameHost. This CL thus adds the infrastructure at the granularity
of WebState itself, with method names making it clear that the requests
are implicitly coming from the main frame.

This CL also adds an EarlGrey test of web_shell that serves as an example
usage of this infrastructure.

Finally, this CL does not replicate //content's mechanism for ensuring
that per-frame interface requests get filtered through service manifest
specifications. Setting up this mechanism necessarily adds indirection
(as all interface requests go down through the service manager), and it
is not clear that there is significant benefit on iOS (where all requests
are coming from the browser process in any case). We can always add this
support later.

Bug: 746447
Change-Id: I54236fc437b1403a363880f9039c43353ea48564
Reviewed-on: https://chromium-review.googlesource.com/582828
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Ken Rockot <rockot@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#489277}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 269aeb0222e38547727acb333ac70db5a45f2996