Enable viewing passwords on iOS by default
Before this CL, viewing passwords on iOS is only possible once enabled by
experimental flags or Finch config. After the CL, the feature is enabled by
default, and can be disabled by Finch or experimental flags. TODOs are added in
the code to remove the kill-switch after the feature spends some time in
stable. But for now the kill switch is left in place in case it is needed
(either via Finch or a small merge).

Bug: 739404
Change-Id: Iab9c4b4c0b768eb8d720c8bc0e651b585b8de7c7
Reviewed-on: https://chromium-review.googlesource.com/568978
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Commit-Queue: Vaclav Brozek <vabr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#489339}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 36f335ed46479eb089cdc3d0a25c405c46970c34