[Payments] Infer if the PR was shown based on the events.
Previously, the way to set it was to call SetShowCalled directly.

Now, this value is inferred in SetEventOccured when the event is either
EVENT_SHOWN or EVENT_SKIPPED_SHOW.

It has the side benefit of forcing the users of JourneyLogger to set
the events.

Bug: 748165
Change-Id: Id41d13bd72ce0bac4a37041846393d1ed602c166
Reviewed-on: https://chromium-review.googlesource.com/583847
Commit-Queue: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Reviewed-by: Mathieu Perreault <mathp@chromium.org>
Reviewed-by: mahmadi <mahmadi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#489504}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 677bd3e5a3f2d39a0acaca9cf9513f68a9b4d6b9