Smooth animation of the fake omnibox moving up
When the fake omnibox is tapped on iPhone, it moves up.
During this move, the width of the omnibox should increase to match the
width of the real omnibox. This increase of width should be done in the
animation time.

Bug: 748507
Change-Id: I233e56bccc9164c920ce537734dbd94e80523eff
Reviewed-on: https://chromium-review.googlesource.com/586712
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Jean-François Geyelin <jif@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#489673}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9976907bcf629ab0c12902ef11af424cf52f2f92