[iOS] Add New Feature Badge to Reading List Tools Menu Item
Add logic to display a new feature badge on the reading list. The
ToolsMenuConfiguration class includes an option to show the badge
on the reading list cell. Although the tools menu is accessed in
multiple places, the reading list cell is only accessed from the
BVC, so that is the only place the option needs to be set.
ToolsMenuViewController reads the configuration option and
animates the new feature badge appropriately.

The new feature badge and number badge occupy the same spot on
the reading list cell, so they should not be shown at the same
time. However, there are circumstances in which it is valid to
show the new feature badge and number badge simultaneously. One
such situation is if the user has added an item to their reading
list without ever having viewed it. Because the number badge already
draws the user's attention to the reading list menu view cell, it
takes precedence over the new feature badge.

Bug: 738164
Change-Id: I847025f2f8dbfc0333f71eb09fecb26c8d172c77
Reviewed-on: https://chromium-review.googlesource.com/570858
Commit-Queue: Cooper Knaak <cooperknaak@google.com>
Reviewed-by: Ed Chin <edchin@chromium.org>
Reviewed-by: Gregory Chatzinoff <gchatz@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#490101}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e4f495cfc5b0fed6002d00ee24123ee01097380c