Skeleton for showing "Show all saved passwords row"
Shows "Show all saved passwords row" for the case when the password was autofilled
and user clicks on the password field for Linux/CrOs/Windows platforms.
The appearance of the row is guarded only buy flag.

No metrics are added in this CL.

TBR=mahmadi@chromium.org

BUG=739343

Review-Url: https://codereview.chromium.org/2971783002
Cr-Original-Commit-Position: refs/heads/master@{#490365}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 04fc429cf5d4c16316402d43c5435c36d09c480f