Update fake omnibox width on rotation
When rotating the device, the omnibox width should be set to the new
width. This width computation needs to take into account the y offset
of the collection and its new size.
So the calculation needs to be started from the collection.

Bug: 750637
Change-Id: I14dc8f878b815f7f32bd33f88691e9b39782e0d4
Reviewed-on: https://chromium-review.googlesource.com/595741
Reviewed-by: Elodie Banel <lod@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#490956}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1ec680becbc61f72eb6ae3e3354bc6fd229b8ce8