[Payments] canMakePayment() rate limit to iframe + top level origin.
Before this patch, if a payment service provider https://checkout.com
was being used in iframes for both https://momsshop.com and
https://popsshop.com, but with different sets of supported payment
methods, then canMakePayment() would be rejected with "NotAllowedError"
due to the rate limit on the iframe origin.

This patch changes the rate limiting to be based on the concatenation of
both iframe and top level origins instead of only iframe origin.

After this patch, if a payment service provider https://checkout.com is
being used in iframes for both https://momsshop.com and
https://bobsshop.com with different sets of supported payment methods,
then canMakePayment() will resolve with "true" or "false" according to
the user state instead of rejecting with "NotAllowedError".

This patch changes behavior for desktop and Android only, because iOS
implementation of PaymentRequest does not support iframes.

Bug: 742589
Change-Id: I08e15c325fc6027e82be7fddc047312235c82f51
Reviewed-on: https://chromium-review.googlesource.com/591747
Reviewed-by: Mathieu Perreault <mathp@chromium.org>
Reviewed-by: mahmadi <mahmadi@chromium.org>
Commit-Queue: Rouslan Solomakhin <rouslan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#490998}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 115f723900c543655bd4b9993d022be8f7bb54ee