[ios clean] Adds background for Tab StatusBar.
PROBLEM:

Currently the status bar background is black and it definitely doesn't look good, 
I think that even a temporal change should improve this. It will give a us a huge UI boost.

Toolbar/Omnibox animation will change both the toolbar and status bar background, 
for this reason I believe that the Toolbar view should include the status bar background

BUT:

What about the Status bar when the Tab Strip is being shown?
What about the Status bar when the Toolbar is placed at the Bottom? 
What about the Toolbar when TabStrip is being shown (We don't want the Toolbar to 
always have that status bar background on top).

LONG TERM SOLUTION?:

I believe the toolbar needs to know about where it's being positioned, just like 
FIP knows how it's being dismissed (so it can animate or not), Toolbar might need 
to know that in order to set their views correctly. 

In the future the Toolbar might extend its background view when positioned on top 
in order to cover the status bar background, that way when the omnibox is selected
we can easily animate it. (I will create a design doc based on these assumptions
when we are working on solving the Toolbar issue.)

THIS CL:

For now TabContainer will add a background to the status bar, this will disappear 
in landscape mode since its constraints are tied to the topLayoutGuide top and 
bottom anchors. I think this is a good starting point and we can add the necessary 
Toolbar "fake status bar background" on follow up Cl's. For now this hugely improves 
how Bling* looks. 

Screenshots
Portrait: 
https://drive.google.com/open?id=0Byo6-Nuda2jgTXJfUkw0ZEVpWTA
TabSwitcher Portrait:
https://drive.google.com/open?id=0Byo6-Nuda2jgWUpydXhuMWZ3U1E
Landscape:
https://drive.google.com/open?id=0Byo6-Nuda2jgSkJpbE1EcG1lbTA
TabSwitcher Landscape:
https://drive.google.com/open?id=0Byo6-Nuda2jgaEMzREwwRGR5NWc



Bug: 
Change-Id: I7271d14fb44b04849983abe7b387aa9c53b38e30
Reviewed-on: https://chromium-review.googlesource.com/607093
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#493061}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dd1c9a2d325577f11fa4e41d65d43ba99c4513c7