[Payment Request] Does not reset url method identifiers if native apps flag is off
This small change makes it so that the url payment method identifiers vector
in PaymentRequest is not reset if WebPaymentNativeApps flag is set to false.
This is important so that tests that do not need the WebPaymentsNativeApps flag
do not incorrectly break.

Bug: 748556
Change-Id: I68c1cc1bef446e8766f25fd7413d135cb8180993
Reviewed-on: https://chromium-review.googlesource.com/609219
Reviewed-by: mahmadi <mahmadi@chromium.org>
Commit-Queue: Randall Raymond <rayraymond@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#493126}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 80a001ccc8803cb3bc4615c740563621b000a1bd