Toast when add ContentSuggestion to Reading List
Using the command option allows the call to be forwarded to the BVC
where the item is added to the Reading List and a toast is displayed.
Adding directly the item to the Reading List as it was done here did
not trigger a toast.

Bug: 753821
Change-Id: I3789a916c4281a0cbde0ef32f443d9a7fcc0c289
Reviewed-on: https://chromium-review.googlesource.com/608238
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Elodie Banel <lod@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#493312}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c9215b63dc020da1b8c7e9e2090307716bbc3084