[Password Manager] Show omnibox icon and anchored prompt once users start typing password
If the password to be save is a generated by Chrome password, the confirmation bubble will be shown (the bubble that a user sees when she submits a form with generated password).

If the password is not a generated one, a save or update bubble will be shown.

BUG=725883

Review-Url: https://codereview.chromium.org/2915763003
Cr-Original-Commit-Position: refs/heads/master@{#493337}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7cf7408af869f5b605a6f35cdd458380c56fb73d