[iOS] Fix bubble_util LeadingDistance RTL Bug
Fix a bug in the LeadingDistance method that caused it to return the
wrong value in right-to-left languages. The method assumed the input
was always in a left-to-right language. The associated unit tests were
calculated incorrectly leading to false positives.

Bug: none
Change-Id: I30d8bd159edf3978f8bf6a1bf409379c9fcbbd83
Reviewed-on: https://chromium-review.googlesource.com/611258
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Gregory Chatzinoff <gchatz@chromium.org>
Commit-Queue: Cooper Knaak <cooperknaak@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#493635}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 26c0f928d0a2ba5cd998e35083ef661bb3c08c7e