[iOS] Present In-Product Help Bubbles on Warm Start
Present in-product help bubbles when Chrome returns to the foreground
in addition to -[BrowserViewController viewDidAppear:]. Users who
don't use the tab switcher or tools menu might never trigger
viewDidAppear, meaning it would only occur on cold start. Presenting
the bubbles when entering the foreground allows many more opportunities
to trigger the bubbles.

However, the bubbles can only be shown if the BVC is currently
visible. If the user backgrounded Chrome with all tabs closed,
reopening Chrome causes a tab to be created. The animation for the
bubble interferes with the animation for creating a new tab, so the
bubble is not presented if there were no tabs and the BVC is
presenting one as part of the transition to the foreground. Once the
new tab animation is complete, viewDidAppear will be called and the
bubble will be presented properly.

The feature engagement tracker prevents multiple bubbles from being
shown at once. If the warm start entry point presents the bubble,
feature_engagement::Tracker->ShouldTriggerHelpUI will return |false|
and not trigger the bubble in viewDidAppear.

Bug: none
Change-Id: I8a56da0e9287490147286c745f22c3041640f931
Reviewed-on: https://chromium-review.googlesource.com/611605
Commit-Queue: Cooper Knaak <cooperknaak@google.com>
Reviewed-by: Gregory Chatzinoff <gchatz@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Rohit Rao (ping after 24h) <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#493853}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e963d670f16de333f777abd6a8369f2a5af1689d