Fix NTP Home Earl Grey tests
The EG rotation does not work on iPad.
And the test case objects are not destroyed at the end of the test case
so the scoped command line needs to be reset at the end of the tests.

Bug: none
Change-Id: I8ada6aa03aa75bca058e3979e341fa34527240da
Reviewed-on: https://chromium-review.googlesource.com/616764
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Jean-François Geyelin <jif@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#494753}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e67e7aaea90131f684b1228d1e41c842e7111dea