Ensure History > Recent Tabs restore preserves window disposition.
This plumbs the window bounds, show state and workspace through the tab restore stack.
This ensure that "Window" entries are restored with their previous window parameters.

This is being submitted with NOPRESUBMIT, because the code uses a deprecated API (Time::To/FromInternal). Unfortunately, it can't be updated to not use it as their exists user data with these values. The code isn't "new", simply refactored.

BUG=718042
NOPRESUBMIT=true

Review-Url: https://codereview.chromium.org/2868983003
Cr-Original-Commit-Position: refs/heads/master@{#495707}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1ba090f0f5fa59c98bd4d8b5c17d6ae7192c749