Reland of "Add EG test for ContentSuggestions context menu"
This CL adds EG tests for the context menu shown when long pressing
a Reading List item in Content Suggestions.
The original CL was reverted because the compilation on iOS was failing
on the StandardResponse. This is because a std::move is needed when
returning a subclass of a class contained in a unique_ptr.

Bug: 749981
Change-Id: I1c9776c2e2f97de297585d3453bee48d2ebd5ca6
Reviewed-on: https://chromium-review.googlesource.com/623067
Reviewed-by: Elodie Banel <lod@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#495893}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 48b152c9556d398a4cd6113b8ba286c7021ca116