Capture snackbarProvider in local var for async blocks.
When code was moved around in http://crrev/c/569985, the code that
shows the snackbox for "Forms filled" was moved from BVC into
ActivityServiceController. A side effect of this move is that the
code that shows the snackbar is now relying on the snackbarProvider
ivar. This ivar is nil' before it is actually used. Thus the block
that displays the snackbar is actually just sending a message to nil.

Bug: 756966
Change-Id: If8e8759bda6713e4b9b39c15d738cd3b5e8f32aa
Reviewed-on: https://chromium-review.googlesource.com/622155
Reviewed-by: Rohit Rao (OOO until 8-30) <rohitrao@chromium.org>
Commit-Queue: Peter Lee <pkl@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#495925}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fb0751e9764ac9f0aac45dbdda65ac878112838b