[Payment Request] Disables Cancel button only once the response is returned
Prior to this patch, the Cancel button was being disabled as soon as the
payment method details were received but before the response was ready.
Preparing the response may require normalizing addresses which is an async
operation. User should be able to still cancel the payment during this
period. After this patch, the Cancel button is disabled only when the
response is ready to be sent off to the merchant.

Bug: 758153
Change-Id: I1c454d48604039f3273b34da7597b246dd2f4c44
Reviewed-on: https://chromium-review.googlesource.com/629836
Reviewed-by: Marc-Antoine Courteau <macourteau@chromium.org>
Commit-Queue: Marc-Antoine Courteau <macourteau@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#496796}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5300b56b4bb4dad39c8044ed559ea803f92bdf48