Fix the toolbar shadow for ContentSuggestions
The bottom toolbar displayed on ContentSuggestions have a shadow if the
content of the collection is displayed below the toolbar.
On iPhone, in order to allow the user to scroll, the content of the
collection is extended to be able to scroll the fake omnibox to the top
of the collection. So collection.contentSize cannot be used directly.

Bug: 757464
Change-Id: Ia8a6bc8d32d38a9461b591e335cc42c99a05f728
Reviewed-on: https://chromium-review.googlesource.com/628880
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Jean-François Geyelin <jif@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#497005}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e3caf09bc1a41c6fbae43a18e0356db1cd7d487