Plumb dispatcher through PasswordController as property.
This CL plumbs the dispatcher through PasswordController as a property
instead of through an initializer. This change better documents that
the dispatcher is held weakly. Since the dispatcher is no longer
required in the initializer of PasswordController, the
PasswordController and all of its dependent tab helpers can be moved
back to attachTabHelpers in Tab.

Bug: none
Change-Id: Icf3313aa01215a400af7e04f3effe808aded9cce
Reviewed-on: https://chromium-review.googlesource.com/630504
Commit-Queue: Gregory Chatzinoff <gchatz@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#497176}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 58806e470b539346e27d09af7b76a79ad4a097dc