Adds string resource to be used by email client app selection title
When user taps on a link that is a mailto:// URL scheme and a default
mail client app has not been set, an action sheet comes up asking user
to make a choice. This is the heading for that action sheet.

Bug: 755524
Change-Id: I33337698655a0cbdc243b1f8c5a03bca69f9ea61
Reviewed-on: https://chromium-review.googlesource.com/634263
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Peter Lee <pkl@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#497240}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8230729c7bf17e20fb1663dd59bc8902d23fb627