DCHECK in PaymentRequestEGTestBase if WebPayments are not enabled.
It is not obious that Payment Tests require WebPayments flag to be
enabled. DCHECKing and explaining how to properly run tests can be
useful for sheriffs.

DCHECK was used instead of GREYFailWithDetails, because it
brings more attention.

Bug: None
Change-Id: Id913cbd10ae450e658c67231c9f1591f2f9b994d
Reviewed-on: https://chromium-review.googlesource.com/636383
Reviewed-by: mahmadi (Moe) <mahmadi@chromium.org>
Reviewed-by: Mike Baxley <baxley@chromium.org>
Commit-Queue: Eugene But <eugenebut@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#497573}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: adac9a93acdd41666d3ec3bceee1e29158ea3308