Implemented WebStateContentIsSecureHtml in credential_manager_util.mm
Implemented function WebStateContentIsSecureHtml for credential manager
and wrote unit tests for it. This is the result of splitting the CL
https://chromium-review.googlesource.com/c/chromium/src/+/631881 and
addressing the comments.

Note that this CL depends on ios/chrome/browser/payments and it will be
modified as soon as
https://chromium-review.googlesource.com/c/chromium/src/+/631881
is landed.

Bug: 435048
Change-Id: I47930e32441be6cea649699b654ccb1d12e615f2
Reviewed-on: https://chromium-review.googlesource.com/638270
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Tomasz Garbus <tgarbus@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#498419}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f8119c8911046893214e97cb28c3710ecee0572