Relayout ContentSuggestions when presented
ContentSuggestions might be changed will not being presented.
For example when the device is rotated while the stack view is
presented. In order to have the view adapted to its new layout, the
layout needs to be recomputed when the view is shown.

Bug: 760085
Change-Id: If0f9e7d5a13b84395d0b6710c0c9b7cc10011155
Reviewed-on: https://chromium-review.googlesource.com/642888
Reviewed-by: Jean-François Geyelin <jif@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#498469}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 78a9901480541606747a20b764f447f449c0bbd9