Adapt scroll offset to content on back navigation
When the user navigates back to Home using the back arrow, the offset
stored might be bigger than the content of the collection.
If the collection is set to an offset bigger than the content, some
constraints are not possible to satisfy and the omnibox is not
displayed.
This CL fixes it by limiting the offset.

Bug: 760439
Change-Id: I5f30bcc4e36d3ff03d9caf809bfd1e0e85cd3d5b
Reviewed-on: https://chromium-review.googlesource.com/645348
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#498800}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e107adb0da7ba682fc76707721d0c558f2c7232f