Test ContentSuggestions opening and back
Add an EG test to ContentSuggestions to test the behavior of the
collection when loading more items, scrolling down, opening an item,
going back.
It tests that the position of the collection is correct and that the
omnibox is correctly displayed, to prevent the bug 760439.

Bug: 760439, 749981
Change-Id: I20efb3c0ecb6799059eab359c47dd5ae763b1965
Reviewed-on: https://chromium-review.googlesource.com/645970
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Jean-François Geyelin <jif@chromium.org>
Reviewed-by: Jan Krcal <jkrcal@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#499477}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6c143cff422710413b2ff7b13da6e17175806352