Delay setting web toolbar controller trailing view hidden state.
When leaving the app in the focused omnibox state and reopening using
the search widget's new search button, the omnibox is closed due to
closing all modal views in the startup sequence, and then re-opened.
Setting the view hidden occurs in the completion block of the animation
which, due to the animation's duration, occurs after setting it back to
visible when re-opening the omnibox. This changes makes it so that the
hidden setting occur in the right order, leaving the view in a correct
final state. However, it does not solve the fact that the omnibox is
hidden then re-opened, which still leads to flickering.

Bug: 754119
Change-Id: I95c785adfdca5ff214b010031b24fde4c475e70e
Reviewed-on: https://chromium-review.googlesource.com/654999
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Elodie Banel <lod@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#500554}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ea473f6717656b6331c9afcec59972ee0a9439fd