Fix call to ReadingListDownloadService::CleanUpFiles.
If the application is shutting down, the task may run on a shutdowned
service.
Make the call static (so it does not need the service) and mark it
SKIP_ON_SHUTDOWN.
Also mark it USER_VISIBLE as it delays the offline feature of ReadingList.

Bug: 762577
Change-Id: I05acf8e58b4dcdf9d21d5ee7e3f352d48f76f699
Reviewed-on: https://chromium-review.googlesource.com/654860
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#500592}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a859436434e4b45c9dacc9e5f0bd56c8616dba9e