Unfocus omnibox when showing context menu on NTP
When the user long press an item on the NTP, a context menu is shown.
The omnibox should be unfocused when this context menu is presented.
It is only necessary to specify it on iPad as showing the alert menu
on iPhone automatically unfocus it.

Bug: 763907
Change-Id: I9d1b8bcf94b64867464ba9ff1ab0a49c423f3a47
Reviewed-on: https://chromium-review.googlesource.com/663342
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#501281}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 53278999f329a7f00e04af2ff61cf391e7cd1354