Hides the ink using dedicated method
To hide the ink on the empty items, ContentSuggestions was using a
transparent color for the ink instead of not starting the animation due
to an MDC bug. As the bug is fixed, the correct method can be used.

Bug: 724493
Change-Id: Id4a63ddcc9977b9c017f882bcf17c2a026ee13b2
Reviewed-on: https://chromium-review.googlesource.com/663340
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#501300}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d57638fdda374318c746e30f29ab35e4b1a7066f