[ios] Uses SadTab box in clean.
This CL implements the Sad Tab functionality for the new architecture.

-Adds a delegate setter to SadTab helper since the helper needs to
outlive the Webstate. Since the delegate could previously only be set
on init. This meant that the delegate had to persist until Webstate 
destruction, which is not ideal since we don't want the delegate to be
a coordinator close to root. For this reason the helper is created
as soon as a Webstate is Added and its delegate is added when the Tab is
being displayed.

Screenshot:
https://drive.google.com/open?id=0Byo6-Nuda2jgeHZhbFBYQTl6Qnc

Bug: 
Change-Id: I42c3c6710233b2148cc3bb696e0017db9c3c205f
Reviewed-on: https://chromium-review.googlesource.com/656518
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#501641}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2573ac31a2c7cc0d7db53f5f5a365e35eeacee46