Fix DCHECK when displaying empty icon in content extension.
When saving an icon in background, the sandbox may already
be locked. This result in saving failure and an empty icon
being used in extension.
This CL
1. Avoid to save empty icon
2. If the widget meets an empty icon at some point, tries
to show a default icon instead of a black one.

Bug: 762517
Change-Id: I66c248f00524900e2c240acb94bc1ed4b5654d5f
Reviewed-on: https://chromium-review.googlesource.com/663261
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Eric Noyau <noyau@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#501928}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8c6dd427eb22d3496cce2d6624df5413b42b8698