[iOS] Move handling of favicon out of Tab.
Remove Tab -favicon getter and change client code to instead fetch
the favicon via the FaviconDriver (with null checks as the driver
may not be installed during unit tests).

Move the code that save the favicon to the NavigationItem from Tab
to WebFaviconDriver (more logical location if the WebFaviconDriver
is responsible for all access to the WebState's favicon).

Move the code that listen to WebState navigation events from Tab to
WebFaviconDriver as it is already a WebStateObserver and this make
the object the one location responsible for fetching and updating
the WebState's favicon.

Remove Tab's FaviconDriverObserver that forwarded the notification
that a new favicon was fetched and instead introduce an observer
in TabModelObserversBridge (as TabModel was forwarding the event
to TabModelObservers anyway).

Remove Tab -fetchFavicon and the call from TabParentingObserver as
it was only useful when restoring a Tab in the background. Instead
move the code into the deserialisation code.

Bug: 760556
Change-Id: I2dc96c62d5d437cd212fafbc6b0cd0612e190ae6
Reviewed-on: https://chromium-review.googlesource.com/663362
Reviewed-by: Rohit Rao (ping after 24h) <rohitrao@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#501938}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7178d4c924a6ecddc3432febadefcd37ea6ec7dd