Add an accessibility label for the fake omnibox
This CL adds a view used only to present an accessible element of the
size of the fake omnibox. Before this CL, only the label was accessible,
allowing the selection of the content visually hidden by the omnibox
using the margin between the label and the border of the fake omnibox.

Bug: 760002
Change-Id: I3ac5f1105709d57290f2a98963434e8fb22457a6
Reviewed-on: https://chromium-review.googlesource.com/660537
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#502218}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 13c981c4f6d5b02826130ffbc1730a580aad64b2