Tapping fake omnibox focuses the real omnibox
Adding a view to have the whole fake omnibox be accessible prevented
interactions with it, as the accessibility view is intercepting the
the touches.
This CL changes the accessibility view to a UIButton, letting it
handle the tap directly.

Bug: 760002, 766455
Change-Id: I34a4297eee16f5811f10a8d8fe0c3b0fa3d6b045
Reviewed-on: https://chromium-review.googlesource.com/671019
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#503115}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b98755f1b057076453fa5914243fc6e2c687ede5