Do not use snapshot when restoring NTP
When restoring the NTP after a cold start, a grey snapshot is briefly
shown as a grey snapshot is shown for all pages displayed after the
WKWebView is flushed, for example after a cold start.
The grey snapshot is faded out once the page is loaded.
As the NTP is a native view, it is not loaded and the snapshot is
directly faded out, creating a flickering experience.
This CL make an exception for the NTP when displaying the snapshot.

Bug: 765737
Test: Try to reload various pages after a cold start. They should have a grey snapshot presented during the loading.
Change-Id: I90462a76012da78291aa0df1d6f66f91afeefb67
Reviewed-on: https://chromium-review.googlesource.com/674869
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Rohit Rao (ping after 24h) <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#503132}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 32fe0dda8215196d9b2e64512eee17c9c283059e