Remove use of parent view from PageInfoVC.
The PageInfoViewController class used to add a subview directly to its
parent view. This CL moves that logic to the presentation provider, so
that the parent view does not need to be passed to the Page Info
component.

Bug: none
Change-Id: I150cae3928adebcb378c7d7ed968a59885ca9401
Reviewed-on: https://chromium-review.googlesource.com/673873
Commit-Queue: Gregory Chatzinoff <gchatz@chromium.org>
Reviewed-by: Rohit Rao (ping after 24h) <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#503230}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b6a01f7921e37e8e5642d087faee79f5dc17ee43