Fix "Navigate back" gesture on bookmarks
The navigate back gesture was broken on bookmarks because opening a new
folder removed the delegate of the gesture recognizer, preventing the
gesture to be triggered.
This CL fixes it by resetting the delegate to the ViewController when
the view is shown.

Bug: 768339,768368
Change-Id: I428bb2838ad7dc1d88d5aa1e21ee74e73493c638
Reviewed-on: https://chromium-review.googlesource.com/681702
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#504072}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 95e7a1505a7ff5f02b534a341e90641ab64aa16a