ios:  Make PrefServiceAdapter use the network pref store.
This means that it only lives on on thread: The network thread, instead
of jumping between the UI thread and the IO thread, since the network
pref store lives on the IO thread, just like the network stack.

A similar change was made to chrome/, as part of network
servicification (Though in Chrome it's created by the NetworkService
now). In a followup CL, I'll simplify HttpServerPropertiesManager to
only support the case where the prefs thread is the network thread.

Bug: 768884
Change-Id: I92c97e8e5fb3f5cab46f1c998b92eaec48a3db5f
Reviewed-on: https://chromium-review.googlesource.com/687803
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#505375}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f077446d3f9ce9f634efbcaf6d760b0cc3bc1e6a