Use the search icon for iOS omnibox calculator matches.
This matches the behavior on Android and avoids a DCHECK when calculator
matches are produced. It would be better if both platforms used the
calculator-specific icon (looks like an equal sign) but that's only
available as a vector graphic. If we can get a png or start using vector
icons on Android and iOS, we should switch. But for now this is an
improvement.

Bug: 754221
Change-Id: Ie589349e6507b7e0b21896df07b4354e4bc37714
Reviewed-on: https://chromium-review.googlesource.com/699315
Reviewed-by: Rohit Rao (ping after 24h) <rohitrao@chromium.org>
Commit-Queue: Justin Donnelly <jdonnelly@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#506382}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7f75f3241a16cbee7f5c72054cc954b343a34cfb