Make CrxInstaller::Install non-blocking.
This is a mechanical change from the callers perspective,
but it affects the execution flow of the component installers and
the implementation of the update client.

The idea here is to make CrxInstaller::Install non-blocking, so that
its callers can defer the execution some of the install tasks, and
notify back the update client when the installation is complete.

Change-Id: Ibeb7b83f30ced4532daa134eae093e3987345b7f
Reviewed-on: https://chromium-review.googlesource.com/701315
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#507209}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f40ab4b3d8b381bcc13cbc7aaed98bef852b9a93