[Nav Experiment] Bug fix for session restoration in WK nav manager.
Moved ExtractSessionState to after CRWWebController creation in the init
sequence of WebStateImpl because WKBasedNavigationManager relies on
CRWWebController to replay session history into the web view. As best as
I can tell, CRWWebController does not depend on the states restored in
ExtractSessionState, so reordering is safe. LegacyNavigationManagerImpl
is agnostic to when CRWWebController is initiated, so it is not affected
by this change either.

Removed extra decodeURIComponent() call in restore_session.html because
the URLSearchParams.get already unescapes the query value before
returning. The extra decodeURIComponent() incorrectly unescapes query
parameters in URLs that are to be restored and can break session restore.

Manually verified that after this change that basica navigation and
history restore functionality is working using WKBasedNavigationManager.

Bug: 734150
Change-Id: I860287a33bc01bafa53b37d479dac2b894813b27
Reviewed-on: https://chromium-review.googlesource.com/705075
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#507771}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f4b3f2450cd73aabdf0c833fb416c5086d9d0c3b