[Payment Request] Show a "Choose" call to action.
When no value is selected, instead of showing the ">" indicator, this
will display "CHOOSE", to indicate to the user what data is missing to
be able to complete the transaction. This matches what is done on
Desktop.

Bug: 
Change-Id: I40d0c2530848f5863e043687bebf2a8d74cdd315
Reviewed-on: https://chromium-review.googlesource.com/703374
Commit-Queue: Marc-Antoine Courteau <macourteau@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#508060}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 87dd49a9ab6415554c3a6dacd7a2fd21295f63f0