Remove |manifest| as a parameter from CrxInstaller::Install
This is a mechanical change.

The change removes the |manifest| parameter from the CrxInstaller::Install.

In the current implementation, the CRX manifest is read once by common
code in the update_client, then it is handed over to implementors of
Install, and implementors of ComponentInstallerPolicy.

The current component installers mostly ignore the manifest data.
With few exceptions (which ignore the manifest data), the component
nstallers are implemented in terms of ComponentInstaller and
ComponentInstallerPolicy. This change refactors how the component
installer reads the manifest.

For future feature work, there are requirements for the manifest to be
read by sandboxed code. Therefore, reading of the manifest must be happen
outside the implementation of CrxInstaller::Install, so that such code can
have the flexibily in how the reading of the manifest occurs.

Bug: 773923
Change-Id: I3393d1c9258b53ebfeec3cebab56429a41f692d0
Reviewed-on: https://chromium-review.googlesource.com/714400
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Joshua Pawlicki <waffles@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Reviewed-by: Bernhard Bauer <bauerb@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Sorin Jianu <sorin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#508799}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7aa6d1f66c9393fa144b370bbe960b5beb50e5de