[ios] Reset StackViewController state after a new tab animation.
The _isActive flag and |clearInternalState| are called indirectly when the stack
view is closed through viewDidDisappear, but it is better to reset state
explicitly when the animations are complete.  This matches the workflow that is
used when an existing Tab is opened from the stack view.

A future CL will switch to VC presentation instead of containment, and that will
result in viewDidDisappear no longer being called in the same way.

BUG=768563

Change-Id: I0c0dd453ca4c49a7f34ff072dc9e52e9fc551c2a
Reviewed-on: https://chromium-review.googlesource.com/723166
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Commit-Queue: Rohit Rao (ping after 24h) <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#509598}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a7727c2b3105ecb5ad5c30e6f8cff174a56f753a