Invalidate layout when NTP will appear
When the NTP will appear, the layout needs to be invalidated. It ensures
that the insets for the Most Visited section are correct when the device
is rotated while another screen (e.g. settings) is presented in front of
the NTP.

Bug: 771149
Change-Id: If8c926fbb0e5cca860b31c4ccc7b672fe49b9e80
Reviewed-on: https://chromium-review.googlesource.com/726082
Reviewed-by: Jean-François Geyelin <jif@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#510069}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 57f53584606e5fc26c3fcb8698f2e0d577b2d20b