[ios] Remove -topPresentedViewController from -showSettings
MainController's -topPresentedViewController is deprecated. It
performs a walk from the bottom of the VC hierarchy to the top
most VC. But this walk is not necessary when the presenting VC
is known.
    
This CL plumbs a baseViewController (either tab switcher or
BVC) to the tools menu so that the presenting VC is known
when calling -showSettings.
    
This CL also plumbs a baseViewController into the 
KeyCommandsProvider.

Bug: 
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: I54bfee925c8b760e3b596d09cf3466506f46e464
Reviewed-on: https://chromium-review.googlesource.com/714648
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#511434}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8e4cfe0345035e34e327b47a916ea1e78e2105d8