[ios] Remove -topPresentedViewController call for showClearBrowsingDataSettings
MainController's -topPresentedViewController is deprecated. It performs a walk
from the bottom of the VC hierarchy to the top most VC. But this walk is
not necessary when the presenting VC is known. In the case of
showClearBrowsingDataSettings, the presentation is made from the
HistoryPanelViewController.

Currently, MainController needs to keep track of the presentation state of
settings. So this CL does not attempt to move the presentation call locally
to HistoryPanelViewController. Instead, the historyPanelViewController is
passed through the dispatcher to the mainController to make the presentation.
In the future, localizing the call would be preferrable.

Bug: 
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: Id340ba37283b4fad5b34ad1aa8e805691c037a79
Reviewed-on: https://chromium-review.googlesource.com/702683
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Rohit Rao (ping after 24h) <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#511456}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8e57ed9610ff9ff4a678775fa680af4a427472bf