Reland [ios] Inserts and deletes history items in the same BatchUpdates block
Original CL at https://codereview.chromium.org/2800473004/

The theory is that performBatchUpdates processes deletion indexes in a
first pass, and then processes insertion indexes. The original CL
created insert indexes first, and then deletion indexes, which is the
suspected reason for needing the rollback. This CL addresses this by
deleting first, and then inserting. The performance ramification is
that we now pay the cost of constructing the array of items on all
query returns, not just when filtering is needed.

The only other change between this and original is that the filtering
method is renamed.

Bug: 707241, 776111
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: I93f0aa9f8d65c43094c1d6ec86eda5565432b4ff
Reviewed-on: https://chromium-review.googlesource.com/730655
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#511895}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a0c9194bc3239a2e3055667504e48238f80a7f23