[iOS] Record which NavigationItems were created for voice searches.
In iOS11, there is a slight delay before a navigation is started, so
Tab's property was being rewritten to NO.  By using
VoiceSearchNavigations, this state is associated with the
NavigationItems themselves.  This ensures that the voice search bar
is shown and that the TTS results are properly played.

This is patch 1 of a 3-sided patch.  It must be landed before
its associated downstream patch and crrev.com/c/740086.

Bug: 773895
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: If24a088006ddc41dd9277a4b04588cac32d06c4d
Reviewed-on: https://chromium-review.googlesource.com/740103
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: Kurt Horimoto <kkhorimoto@chromium.org>
Commit-Queue: Kurt Horimoto <kkhorimoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#512059}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 208a1e87a413d81379e1d8c607114621592d6c21