Implement HTTPBad Phase 2 for iOS
This change updates the iOS HTTPBad implementation to more closely match
the implementation in /content. Specifically, the legacy bitflags are
moved from the ContentStatusFlags to a SSLUserData object attached to
the SSLStatus, fixing a layering violation. Updates to the flags are
handled at the browser layer (rather than in the /web layer).

A WebStateObserver watches for the FormActivityRegistered event and
sets the insecure_input_event flag on the SSLUserData if the page is
loaded from a non-secure context, implementing HTTPBad Phase 2 for
iOS.

Bug: 720094, 689082
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet
Change-Id: I2f54a3aff212f8378a52b809c717a577ca4c8a69
Reviewed-on: https://chromium-review.googlesource.com/755889
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Commit-Queue: Eric Lawrence <elawrence@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#515075}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e5cdce168c1dd9490d6f4cc4087d0219bd6817f3