Add delay on EG test
This CL adds a 0.5s delay on the ContentSuggestions tests to allow the
New Tab Opened animation to complete. This is needed as the animation
is a CAAnimations, EG has trouble synchronizing on it.
Using a loop to check if the first tab has came back does not work as
the steps inside the loop is too short for the animation to have time
to continue. Using it froze the animation.

Bug: 783169
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I71e247d9e373df96f601e613eb0bec83704ac09a
Reviewed-on: https://chromium-review.googlesource.com/763408
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#515525}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a7bb9e7c1e3b3715334654540915d5d89d8d692e