[iOS] Update Edit button after password deletion
Chrome settings on iOS allow selecting a set of saved password entries
and deleting them. To select the entries, the view enters the "Edit"
mode. If there are any entries left after deletion, the view stays in
"Edit", but if there are none, it should go back to the default mode
and show a disabled "Edit" button.

The code clearly intended to do that, but called |setEdit| by mistake
on the view controller instead of on its 'editor' property. This CL
fixes that.

The CL also fixes the glitch that Edit was not disabled if the last
password was deleted from the detailed view.

Bug: 755108
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: Ib28679ecf0502f3231ea7d918acd576ce048489e
Reviewed-on: https://chromium-review.googlesource.com/774745
Commit-Queue: Vaclav Brozek <vabr@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#517109}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7eb7c274db08e9fb1cea558f9f6e9240dc23cf9b