Improve toolbar snapshotting and layout.
Currently, on iPhone X, if you open the tab switcher, rotate to another
orientation and click on a tab, there's a DCHECK during snapshotting.

This resolves the DCHECK and also removes the need for non-required
constraints in the toolbar. Having only required constraints in the
toolbar allows for easier autolayout inside the toolbar.

The DCHECK resolution is as follows: currently prior to snapshotting
the toolbar, we try to resize it to its future size after rotation,
which is actually ignored because it's contained in BVC's view that
is not updated to the post-rotation frame yet.
This change makes the BVC's view temporarily take the new frame for
snapshotting, then return to the original size once the snapshot is
captured. This is OK because BVC is not in a window at that moment.

Bug: 
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I75909e6170ba52681e0d1a3b385060a7fd7e85d7
Reviewed-on: https://chromium-review.googlesource.com/776664
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#518636}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a729577083660ca4d61cc2c9b660a45e063b2fa8