Allow word-wrap of button labels in iOS Bookmarks' bottom toolbar
This is a follow-up CL of crrev.com/c/780705. In which the heightAnchor
constraint doesn't work on iOS 9, 10 so that word-wrapping of button
labels doesn't happen when needed.  This CL manually checks if word
wrapping is needed and adjust the height of toolbar accordingly (the way
reading_list_toolbar is using).

Bug: 778979
Change-Id: I9fa2b46ca18fc4cc44d6c75c190df7c88304c180
Reviewed-on: https://chromium-review.googlesource.com/788633
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Marti Wong <martiw@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#519170}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b30c10599e62a282e5e33c0c0f124088fe600ded