[iOS][Signin] Adding SigninPromoViewMediator.signinInProgress property
To address [1], I suggest to remove |SigninStarted|. This will be more
obvious that the enum is only related to the state of the Sign-in
promo view, and nothing related to the sign-in interaction controller.
With this, SigninPromoViewMediator.signinPromoViewState can switch
from |Unused| to |UsedAtLeastOnce|, as soon as the primary and
secondary buttons are clicked.

The information about if the sign-in interaction controller is shown
or not, is now saved into SigninPromoViewMediator.signinInProgress.

This change help to make avoid UI glitches by not refreshing the
sign-in promo view as long as the sign-in interaction controller is
displayed.

[1]: https://chromium-review.googlesource.com/c/chromium/src/+/788378/5/ios/chrome/browser/ui/authentication/signin_promo_view_mediator.h#31

Bug: 
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: Iecafbc32ad4893e0ff44d891022157b9727514e1
Reviewed-on: https://chromium-review.googlesource.com/789078
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#519333}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f026f770833745dd26bdc3f86b67c3ba9b5648fa