[iOS] Fix stack view card sizing for BVC presentation.
When the BVC is presented, the timing of when the stack view's
-viewWillAppear: method call is changed such that it doesn't yet
reflect any size changes to the stack view since the last time it ran.
(For example, if the device is rotated between seeing the stack view).

This means that the card sizes will be computed based on an incorrect
size for the stack view.

This CL fixes this for non-iPhone X devices by (in MainPresentingVC)
calling the new -prepareForDisplayAtSize: TabSwitcher method before
-dismissViewController is called, which kicks off the UIKit-managed
transition process.

-prepareForDisplayAtSize: is a no-op in the iPad TabSwitcher; in the
StackViewController it just updates the view bounds to have the passed
size if they don't already. This means the right size will be set when
-viewWillAppear: is later called.

In order for this to work correctly, MainPresentingVC also needs to
make sure its child view controllers are properly constrained, so that
any size changes to MainPresentingVC's view (which happens in the
transition animator) propagate to its child view controller.

This CL also includes some assorted cleanup in StackView-related files:
 - StackViewController's -loadView method is changed to -viewDidLoad.
 - CardSet's _view ivar is renamed to match the property it backs.
 - The iOS8 guards around IsPortrait() are removed, since the minimum
   OS version is now 10.0.

Bug: 787832
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I1f998f37aa14a1ce12512d912a49b0b274d3043a
Reviewed-on: https://chromium-review.googlesource.com/800870
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#520583}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1b37326136d3d9891ea11af23c0293fe78f5b988