[ios] Prevents omnibox to shrink on iPad CleanToolbar.
On ipad, the omnibox creates a right margin so WTC can layout the bookmark and voice buttons
on top of it.
This is not the case on the clean toolbar, so this CL disables the margin creation when the omnibox
is being used by the clean toolbar.

Screenshot:
https://drive.google.com/open?id=1dX-UmlHJstycs_jBAcGsbkcrm3JQxZfh

Bug: 791967
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: Ifd3ab1cf9baf09c663c4ab3a6f05fd544c190eee
Reviewed-on: https://chromium-review.googlesource.com/813181
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#522548}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 792b2927b474964a9ceebabf9dd35b4e0b0b8b2d