[ios] Add comment to remove -rootViewController
The use of -rootViewController is fine here because it
is used for safe area sizing rather than presentation.
This is necessary because the view is not in the
hierarchy at the time the sizing is done.
We should find a way to completely remove
rootViewController, which may involve moving
when the sizing is done in this case.


Bug: 791784
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: Id69b668a9b08c85b0e0ff66a3dd90ac56f580ecd
Reviewed-on: https://chromium-review.googlesource.com/824432
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#523795}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c14516229d625541261f1d9b6bc290f88d87a2e9