Fix the new bookmark folder name issue on iPad when keyboard dismiss.
Fix the issue that when editing the new folder name on iPad and tap on
the 'hide keyboard' button, the folder name is supposed to commit but
doesn't.  An egtest is added to prevent this bug from happening again.

Bug: 794155
Change-Id: Ie0ef4c1af7f87c55051876bd08331c5d0a0e40a8
Reviewed-on: https://chromium-review.googlesource.com/826585
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Marti Wong <martiw@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#524058}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 54857f4351ae59436f7a7363204b6b6bca309140