[iOS][Signin] Fixing DCHECK and UI glitch in Other Devices tab
When the user taps on the "Continue as ..." button in the sign-in
promo, of the Other Devices tab, the user is signed in at the
beginning of the sync confirmation animation.
Therefore the TabSwitcherModel receives the sign-in notification and
send a notification to the TabSwitcherController to reload before the
end of the animation. While reloading the TabSwitcherController
removes the current TabSwitcherPanelOverlayView that contains the
sign-in promo view and its SigninPromoViewMediator.
The issue is that SigninPromoViewMediator is supposed to live until
the end of the sign-in to handle correctly the metrics (histograms
and user actions).

To avoid that, the SigninPromoViewMediator is now owned by the
TabSwitcherController. And TabSwitcherModel asks its delegate if a
sign-in is in progress. In that case, the sign-in notification is
ignored.
Therefore the UI is not updated while the sign-in interaction is
appearing (see crbug.com/725925)

Once the sign-in is finished (cancel or success), the
TabSwitcherController is notified by SigninPromoViewMediator and then
the TabSwitcherModel can be reloaded.

Bug: 795895, 725925
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: Iefff5975bc251ea39cb06c5d73a4adc135ce317b
Reviewed-on: https://chromium-review.googlesource.com/832391
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#525309}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e273042d34106a89730915faea6bc4f8e0e103bc