Wait until translate infobar is presented.
The method -simulateTranslationFromSpanishToEnglish was causing
flakyness of TranslateTestCase/test{,Incognito}TranslateInfoBar
as the presentation of the infobar is animated. The animation
is implemented in C++ so EarlGrey cannot synchronize with it.

Instead add a method to wait for a matcher to return an element
that is sufficiently visible and call it from flaky method to
ensure that the element is visible.

Use grey_sufficientlyVisible() instead of grey_notNil() in the
test as the latter match elements that are outside of the window
but grey_tap() cannot tap on them.

Re-enable the test as the flakyness has been fixed.

Bug: 798063
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I1047e2d4780e2f999b51ce3df9f3eaf7daf3e77b
Reviewed-on: https://chromium-review.googlesource.com/846746
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#526378}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eddf23e9ede4271428fbe6e45fce7b31438995a3