Reland of  "Fix URL clipping on iOS 11."
This reverts CL 779191.

Wraps omnibox textfield into a container that positions the textfield
inside of itself and clips its bounds so that the optimal clipping is
used. The clipping logic replicates that of the pre-iOS 11 solution
with UITextField overrides that are now broken in iOS 11.
To mask the clipped ends of the omnibox, a ClippingMaskView is used
as a maskView of the clipping view.
All of this is behind kClippingTextfield flag that is enabled by
default.

Bug: 749788
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: I4cc6673200a8fbdc57451dc9120dfd4eeec5a853
Reviewed-on: https://chromium-review.googlesource.com/844079
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#526693}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9234e9ce80dabba5cf04a295a18efb0df6cc5db9