Add timestamp to requests for experimental contextual suggestions.
The experimental contextual suggestions service provides 0-suggest
autocomplete suggestions based on the URL that is visited in the active
tab, at the time the address bar is focused. This CL adds to the request
a timestamp corresponding to the time the user started navigating the
URL for which the suggestions are requested.

This change is a requirement for the process of enabling server-side
logging of the contextual suggestions that are sent to Chrome clients.
If logged, contextual suggestions are to be removed as soon as a
synced user removes from their history the page that was visited at the
time the suggestions were displayed. The timestamp is necessary for
matching suggestions with the corresponding navigation.

TBR=rohitrao
for mechanical changes to ios/c/b/a/*

Bug: 692471
Change-Id: I6939d108cfab91b69640acdbb3b8307687999320
Reviewed-on: https://chromium-review.googlesource.com/879167
Commit-Queue: Gheorghe Comanici <gcomanici@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#531999}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e76e0e99dc3b8b2b7ad502603331ea7104745a1a