Fix OmniboxPerfTest.TestTypeOneCharInTextField
This CL fixes the OmniboxPerfTest.TestTypeOneCharInTextField test which
was failing because it was using the WebToolbarController even if the
flag to disable it is turned on. It was creating a conflict.

Bug: 806685
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: If846ac3ecf0963700da4e1dd71156037c87653c4
Reviewed-on: https://chromium-review.googlesource.com/891018
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#532385}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 429c46198793b67075edcb8cf2ea940e67ada843