[NTP Snippets] Remove usage of internal signin classes in factories
This CL eliminates the minimal usage of SigninManager in the
ContentSuggestionsService factories on Android and iOS, replacing that
usage with equivalent usage of IdentityManager. It also eliminates
the dependencies on SigninManager and ProfileOAuth2TokenService
entirely.

To test the minimal usage change, I verified that the first remote
snippets fetch done on Android on cold start of the browser was
authenticated if the user was signed in, and unauthenticated if the user
was not signed in.

TBR=gambard@chromium.org

Bug: 798413
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: Ifb79778ef385441634e7624a1c99d366a579b0b2
Reviewed-on: https://chromium-review.googlesource.com/904043
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Jan Krcal <jkrcal@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#536352}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe56cf921b0b202657ef1d5bfbb2b25de4f44019