[iOS] Mitigate white screen while uploading crash reports.
Between when the root view controller is created and the new (or active)
tab is displayed, there is a brief interval when the root view controller
is displayed. This is usually less than one display refresh cycle, but
if (for example) a crash is being uploaded on a slow device, it may
be longer than that. If so, a white screen will appear (this is the root
view controller's view when it doesn't contain a tab switcher).

To mitigate this effect, this CL has the main view controller add a copy
of the launch screen view to its view hierarchy. It is removed as soon as
another view controller is added as a child of the main view controller.

This is the same approach that's used to solve a similar issue when showing
the FRE on some devices. It's terrible and we will find something better
soon, but this will resolve the issue for M65.

Bug: 803758
Cq-Include-Trybots: master.tryserver.chromium.mac:ios-simulator-cronet;master.tryserver.chromium.mac:ios-simulator-full-configs
Change-Id: Ife105aa253019b41eac5002a33e69da71eb51944
Reviewed-on: https://chromium-review.googlesource.com/920521
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#537438}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3fb3face370c1039e96da2fdec80869e314eadf0